requirements.txt
fastapi
uvicorn
pandas
numpy
pytz
tzdata
typing-extensions
sniffio
six
idna
h11
click
annotated-types
python-dateutil
pydantic-core
anyio
starlette
scikit-learn
geopy
google-generativeai